+++ tests/libtest/lib1560.c
@@ -73,7 +73,7 @@ static int checkparts(CURLU *u, const char *in, const char *wanted,
       curl_msnprintf(bufp, len, "%s%s", buf[0]?" | ":"", p);
     }
     else
-      curl_msnprintf(bufp, len, "%s[%d]", buf[0]?" | ":"", (int)rc);
+      curl_msnprintf(bufp, len, "%s[%d]", buf[0]?" | ":"", rc);
 
     n = strlen(bufp);
     bufp += n;
@@ -1198,7 +1198,7 @@ static CURLUcode updateurl(CURLU *u, const char *cmd, unsigned int setflags)
         CURLUPart what = part2id(part);
 #if 0
         /* for debugging this */
-        curl_mfprintf(stderr, "%s = \"%s\" [%d]\n", part, value, (int)what);
+        curl_mfprintf(stderr, "%s = \"%s\" [%d]\n", part, value, what);
 #endif
         if(what > CURLUPART_ZONEID)
           curl_mfprintf(stderr, "UNKNOWN part '%s'\n", part);
@@ -1395,7 +1395,7 @@ static int set_url(void)
       if(rc) {
         curl_mfprintf(stderr, "%s:%d Set URL %s returned %d (%s)\n",
                       __FILE__, __LINE__, set_url_list[i].set,
-                      (int)rc, curl_url_strerror(rc));
+                      rc, curl_url_strerror(rc));
         error++;
       }
       else {
@@ -1403,7 +1403,7 @@ static int set_url(void)
         rc = curl_url_get(urlp, CURLUPART_URL, &url, 0);
         if(rc) {
           curl_mfprintf(stderr, "%s:%d Get URL returned %d (%s)\n",
-                        __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                        __FILE__, __LINE__, rc, curl_url_strerror(rc));
           error++;
         }
         else {
@@ -1416,7 +1416,7 @@ static int set_url(void)
     }
     else if(rc != set_url_list[i].ucode) {
       curl_mfprintf(stderr, "Set URL\nin: %s\nreturned %d (expected %d)\n",
-                    set_url_list[i].in, (int)rc, set_url_list[i].ucode);
+                    set_url_list[i].in, rc, set_url_list[i].ucode);
       error++;
     }
     curl_url_cleanup(urlp);
@@ -1455,7 +1455,7 @@ static int setget_parts(bool has_utf8)
         if(uc != setget_parts_list[i].pcode) {
           curl_mfprintf(stderr,
                         "updateurl\nin: %s\nreturned %d (expected %d)\n",
-                        setget_parts_list[i].set, (int)uc,
+                        setget_parts_list[i].set, uc,
                         setget_parts_list[i].pcode);
           error++;
         }
@@ -1470,7 +1470,7 @@ static int setget_parts(bool has_utf8)
       }
       else if(rc != CURLUE_OK) {
         curl_mfprintf(stderr, "Set parts\nin: %s\nreturned %d (expected %d)\n",
-                      setget_parts_list[i].in, (int)rc, 0);
+                      setget_parts_list[i].in, rc, 0);
         error++;
       }
     }
@@ -1503,7 +1503,7 @@ static int set_parts(void)
 
       if(uc != set_parts_list[i].pcode) {
         curl_mfprintf(stderr, "updateurl\nin: %s\nreturned %d (expected %d)\n",
-                      set_parts_list[i].set, (int)uc, set_parts_list[i].pcode);
+                      set_parts_list[i].set, uc, set_parts_list[i].pcode);
         error++;
       }
       if(!uc) {
@@ -1512,7 +1512,7 @@ static int set_parts(void)
 
         if(rc) {
           curl_mfprintf(stderr, "%s:%d Get URL returned %d (%s)\n",
-                        __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                        __FILE__, __LINE__, rc, curl_url_strerror(rc));
           error++;
         }
         else if(checkurl(set_parts_list[i].in, url, set_parts_list[i].out)) {
@@ -1523,7 +1523,7 @@ static int set_parts(void)
     }
     else if(rc != set_parts_list[i].ucode) {
       curl_mfprintf(stderr, "Set parts\nin: %s\nreturned %d (expected %d)\n",
-                    set_parts_list[i].in, (int)rc, set_parts_list[i].ucode);
+                    set_parts_list[i].in, rc, set_parts_list[i].ucode);
       error++;
     }
     curl_url_cleanup(urlp);
@@ -1552,7 +1552,7 @@ static int get_url(bool has_utf8)
 
         if(rc) {
           curl_mfprintf(stderr, "%s:%d returned %d (%s). URL: '%s'\n",
-                        __FILE__, __LINE__, (int)rc, curl_url_strerror(rc),
+                        __FILE__, __LINE__, rc, curl_url_strerror(rc),
                         get_url_list[i].in);
           error++;
         }
@@ -1565,7 +1565,7 @@ static int get_url(bool has_utf8)
       }
       if(rc != get_url_list[i].ucode) {
         curl_mfprintf(stderr, "Get URL\nin: %s\nreturned %d (expected %d)\n",
-                      get_url_list[i].in, (int)rc, get_url_list[i].ucode);
+                      get_url_list[i].in, rc, get_url_list[i].ucode);
         error++;
       }
     }
@@ -1592,7 +1592,7 @@ static int get_parts(bool has_utf8)
                         get_parts_list[i].urlflags);
       if(rc != get_parts_list[i].ucode) {
         curl_mfprintf(stderr, "Get parts\nin: %s\nreturned %d (expected %d)\n",
-                      get_parts_list[i].in, (int)rc, get_parts_list[i].ucode);
+                      get_parts_list[i].in, rc, get_parts_list[i].ucode);
         error++;
       }
       else if(get_parts_list[i].ucode) {
@@ -1649,7 +1649,7 @@ static int append(void)
       ;
     else if(rc != append_list[i].ucode) {
       curl_mfprintf(stderr, "Append\nin: %s\nreturned %d (expected %d)\n",
-                    append_list[i].in, (int)rc, append_list[i].ucode);
+                    append_list[i].in, rc, append_list[i].ucode);
       error++;
     }
     else if(append_list[i].ucode) {
@@ -1660,7 +1660,7 @@ static int append(void)
       rc = curl_url_get(urlp, CURLUPART_URL, &url, 0);
       if(rc) {
         curl_mfprintf(stderr, "%s:%d Get URL returned %d (%s)\n",
-                      __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                      __FILE__, __LINE__, rc, curl_url_strerror(rc));
         error++;
       }
       else {
@@ -1686,7 +1686,7 @@ static int scopeid(void)
                     "https://[fe80::20c:29ff:fe9c:409b%25eth0]/hello.html", 0);
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr, "%s:%d curl_url_set returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
 
@@ -1694,7 +1694,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_HOST returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1705,7 +1705,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_set CURLUPART_HOST returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
 
@@ -1713,7 +1713,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_URL returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1724,7 +1724,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_set CURLUPART_HOST returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
 
@@ -1732,7 +1732,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_URL returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1744,7 +1744,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_set CURLUPART_HOST returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
 
@@ -1752,7 +1752,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_URL returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1763,7 +1763,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_HOST returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1774,7 +1774,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_ZONEID returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1785,7 +1785,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_set CURLUPART_ZONEID returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
 
@@ -1793,7 +1793,7 @@ static int scopeid(void)
   if(rc != CURLUE_OK) {
     curl_mfprintf(stderr,
                   "%s:%d curl_url_get CURLUPART_URL returned %d (%s)\n",
-                  __FILE__, __LINE__, (int)rc, curl_url_strerror(rc));
+                  __FILE__, __LINE__, rc, curl_url_strerror(rc));
     error++;
   }
   else {
@@ -1848,7 +1848,7 @@ static int get_nothing(void)
 
     rc = curl_url_get(u, CURLUPART_ZONEID, &p, 0);
     if(rc != CURLUE_NO_ZONEID)
-      curl_mfprintf(stderr, "unexpected return code %u on line %u\n", (int)rc,
+      curl_mfprintf(stderr, "unexpected return code %u on line %u\n", rc,
                     __LINE__);
 
     curl_url_cleanup(u);
