+++ tests/libtest/lib677.c
@@ -87,7 +87,7 @@ static CURLcode test_lib677(const char *URL)
         }
         else if(ec) {
           curl_mfprintf(stderr, "curl_easy_send() failed, with code %d (%s)\n",
-                        (int)ec, curl_easy_strerror(ec));
+                        ec, curl_easy_strerror(ec));
           res = ec;
           goto test_cleanup;
         }
@@ -108,7 +108,7 @@ static CURLcode test_lib677(const char *URL)
         }
         else if(ec) {
           curl_mfprintf(stderr, "curl_easy_recv() failed, with code %d (%s)\n",
-                        (int)ec, curl_easy_strerror(ec));
+                        ec, curl_easy_strerror(ec));
           res = ec;
           goto test_cleanup;
         }
